feat: merge CustomIcon into just Icon
