+++ src/icons/Icon.tsx
@@ -1,17 +1,29 @@
 import React from "react"
 import { IconDefinition } from "@fortawesome/fontawesome-svg-core"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import { CommonIconProps, iconClassNames } from "./CommonIconProps"
 
 import "./Icon.scss"
 
-export interface IconProps extends CommonIconProps {
+export interface IconProps {
+  /** Element ID */
+  id?: string
+  /** Additional CSS classes */
+  className?: string
+  /** Specify a specific preset size */
+  size?: "utility-sm" | "utility-md" | "badge-lg" | "badge-xl"
+  "aria-hidden"?: boolean
+  "aria-label"?: string
+  "aria-labelledby"?: string
+  tabIndex?: number
   /** Icon SVG metadata imported from Font Awesome */
-  icon: IconDefinition
+  icon?: IconDefinition
+  /** Custom SVG in JSX if not using Font Awesome */
+  children?: React.ReactNode
 }
 
 const Icon = (props: IconProps) => {
-  const classNames = iconClassNames(props)
+  const classNames = ["icon"]
+  if (props.className) classNames.push(props.className)
   const isHidden = !(props["aria-hidden"] === false)
 
   return (
@@ -22,8 +34,9 @@ const Icon = (props: IconProps) => {
       aria-hidden={isHidden}
       aria-label={props["aria-label"]}
       aria-labelledby={props["aria-labelledby"]}
+      tabIndex={props.tabIndex}
     >
-      <FontAwesomeIcon icon={props.icon} />
+      {props.icon ? <FontAwesomeIcon icon={props.icon} /> : props.children}
     </span>
   )
 }
