--- src/icons/__stories__/CustomIcon.docs.mdx
@@ -1,45 +1,0 @@
-import { ArgsTable, Canvas, Story } from "@storybook/addon-docs"
-import { Link } from "../../../documentation/components/Link"
-import CustomIcon from "../CustomIcon"
-
-# Custom Icon
-
-Displays an icon via the supplied SVG markup.
-
-<Canvas>
-  <Story id="icons-custom-icon--custom-icon-using-svg" />
-</Canvas>
-
-## Props
-
-<ArgsTable of={CustomIcon} />
-
-## Best Practices
-
-### When to Use
-
-- Use this component to add visual cues and enhance the user experience in interfaces such as buttons, navigation, and other interactive elements.
-
-### When Not to Use
-
-- Do not use this component for non-iconographic content, such as images or illustrations.
-
-### Usability
-
-- Use icons that are easily recognizable and have a clear meaning. Avoid using icons that are abstract or metaphorical.
-- Use the correct icon according to the context, meaning, and action associated with the element.
-
-### Content Guidelines
-
-- Use only for specific needs that are not covered by the Font Awesome icon set. Otherwise, use the Icon component instead.
-
-### Accessibility Notes
-
-- Use the `aria-hidden` attribute for purely decorative icons that do not convey any information.
-- Provide a text alternative for icons that convey information, such as using an `aria-label` or `aria-labelledby` attribute.
-- Ensure that the icons are visible and can be easily distinguished by users with low vision or color blindness.
-- Avoid using only icons to convey information, as it can be difficult for users who rely on assistive technologies to understand the meaning of the icons.
-
-## Theme Variables
-
-See the <Link href="/?path=/docs/icons-icon--icons">Icon</Link> documentation for a list of variables.
